+++ libs/core/src/controllers/dynamic-styles-controller.ts
@@ -31,14 +31,13 @@ export class DynamicStylesController implements ReactiveController {
    * @param styles - The CSSResult to inject.
    */
   inject(key: string, styles: CSSResult) {
-    const cssText = Array.isArray(styles)
-      ? styles.map((style) => style.toString()).join('')
-      : styles.toString()
-
     if (this.#useLegacyStylesheets) {
+      const cssText = Array.isArray(styles)
+        ? styles.map((style) => style.toString()).join('')
+        : styles.toString()
       this.#applyStylesLegacy(key, cssText)
     } else {
-      this.#applyStyles(key, cssText)
+      styles.styleSheet && this.#applyStyles(key, styles.styleSheet)
     }
   }
 
@@ -73,17 +72,11 @@ export class DynamicStylesController implements ReactiveController {
     this.host.shadowRoot?.appendChild(styleEl)
   }
 
-  #applyStyles(key: string, cssText: string) {
+  #applyStyles(key: string, styleSheet: CSSStyleSheet) {
     if (!this.host.shadowRoot) return
 
-    let styleSheet = this.#styleSheets.get(key)
+    this.#styleSheets.set(key, styleSheet)
 
-    if (!styleSheet) {
-      styleSheet = new CSSStyleSheet()
-      this.#styleSheets.set(key, styleSheet)
-    }
-
-    styleSheet.replaceSync(cssText)
     this.host.shadowRoot.adoptedStyleSheets = [
       ...this.#initialStyleSheets,
       ...Array.from(this.#styleSheets.values()),